Delete item from custom list works and using LoDash made sure no new list was added depending on casing in url

+++ app.js
@@ -3,7 +3,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require("mongoose");
-const date = require(__dirname + "/date.js");
+const _ = require('lodash');
 
 const app = express();
 
@@ -61,8 +61,7 @@ app.get("/", (req, res) => {
 });
 
 app.get("/:customListName", (req, res) => {
-  const customListName = req.params.customListName;
-
+  const customListName = _.capitalize(req.params.customListName);
   List.findOne({name: customListName}, (err, foundList) => {
     if (!err) {
       if (!foundList) {
@@ -104,22 +103,28 @@ app.post("/", function(req, res){
       }
     });
   }
-
-  
-
 });
 
 app.post("/delete", (req, res) => {
 
   const checkedItemId = req.body.checkbox;
+  const listName = req.body.listName;
 
-  Item.findByIdAndRemove(checkedItemId, (err) => {
-    if (!err) {
-      console.log("Item has been removed")
-      res.redirect("/");
-    }
-  })
-})
+  if (listName === "Today") {
+    Item.findByIdAndRemove(checkedItemId, (err) => {
+      if (!err) {
+        console.log("Item has been removed")
+        res.redirect("/");
+      }
+    });
+  } else {
+    List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedItemId}}}, (err, foundList) => {
+      if (!err) {
+        res.redirect("/" + listName);
+      }
+    });
+  }
+});
 
 
 app.get("/work", function(req,res){
